Preserve activation dtype in MLX group-16 affine kernel

+++ gptqmodel/nn_modules/qlinear/mlx_group16.py
@@ -38,4 +38,4 @@ class MlxGroup16Linear(nn.Module):
         result = even + odd
         if "bias" in self:
             result += self.bias
-        return result
+        return result.astype(x.dtype)
